+++ packages/core/src/ruleset/validation/__tests__/validation.test.ts
@@ -41,13 +41,18 @@ describe('JS Ruleset Validation', () => {
     expect(assertValidRuleset.bind(null, invalidRuleset)).toThrowAggregateError(
       new AggregateError([
         new RulesetValidationError('the rule must have at least "given" and "then" properties', [
-          '/rules/no-given-no-then',
+          'rules',
+          'no-given-no-then',
         ]),
         new RulesetValidationError('allowed types are "style" and "validation"', [
-          '#/rules/rule-with-invalid-enum/type',
+          'rules',
+          'rule-with-invalid-enum',
+          'type',
         ]),
         new RulesetValidationError('the value has to be one of: 0, 1, 2, 3 or "error", "warn", "info", "hint", "off"', [
-          '#/rules/rule-with-invalid-enum/severity',
+          'rules',
+          'rule-with-invalid-enum',
+          'severity',
         ]),
       ]),
     );
@@ -365,11 +370,12 @@ describe('JS Ruleset Validation', () => {
       },
     );
 
-    it.each(['#Info', '#i', '#Info.contact', '#Info[*]'])('recognizes %s as a valid value of an alias', value => {
+    it.each(['#Info', '#Info.contact', '#Info[*]'])('recognizes %s as a valid value of an alias', value => {
       expect(
         assertValidRuleset.bind(null, {
           rules: {},
           aliases: {
+            Info: ['$'],
             alias: [value],
           },
         }),
@@ -427,7 +433,7 @@ describe('JS Ruleset Validation', () => {
         [
           new RulesetValidationError(
             'must be a valid JSON Path expression or a reference to the existing Alias optionally paired with a JSON Path expression subset',
-            ['aliases', 'PathItem'],
+            ['aliases', 'PathItem', '0'],
           ),
         ],
       ],
@@ -436,7 +442,7 @@ describe('JS Ruleset Validation', () => {
         [
           new RulesetValidationError(
             'must be a valid JSON Path expression or a reference to the existing Alias optionally paired with a JSON Path expression subset',
-            ['aliases', 'PathItem'],
+            ['aliases', 'PathItem', '0'],
           ),
         ],
       ],
@@ -446,7 +452,7 @@ describe('JS Ruleset Validation', () => {
         [
           new RulesetValidationError(
             'must be a valid JSON Path expression or a reference to the existing Alias optionally paired with a JSON Path expression subset',
-            ['aliases', 'PathItem'],
+            ['aliases', 'PathItem', '0'],
           ),
         ],
       ],
@@ -472,7 +478,10 @@ describe('JS Ruleset Validation', () => {
           }),
         ).toThrowAggregateError(
           new AggregateError([
-            new RulesetValidationError('targets must be present and have at least a single alias definition', []),
+            new RulesetValidationError('targets must be present and have at least a single alias definition', [
+              'aliases',
+              'alias',
+            ]),
           ]),
         );
       });
@@ -498,11 +507,26 @@ describe('JS Ruleset Validation', () => {
         },
       );
 
-      it.each(['#Info', '#i', '#Info.contact', '#Info[*]'])('recognizes %s as a valid value of an alias', value => {
+      it.each(['#Info', '#Info.contact', '#Info[*]'])('recognizes %s as a valid value of an alias', value => {
         expect(
           assertValidRuleset.bind(null, {
-            rules: {},
+            rules: {
+              a: {
+                given: '#alias',
+                then: {
+                  function: truthy,
+                },
+              },
+            },
             aliases: {
+              Info: {
+                targets: [
+                  {
+                    formats: [formatA],
+                    given: ['$'],
+                  },
+                ],
+              },
               alias: {
                 targets: [
                   {
@@ -624,7 +648,7 @@ describe('JS Ruleset Validation', () => {
                 targets: [
                   {
                     formats: [formatA],
-                    given: ['#.definitions[*]'],
+                    given: ['$.definitions[*]'],
                   },
                   {
                     formats: [formatA, formatB],